changed price to decimal

+++ backend/db/migrations/20220630230354-create-room.js
@@ -50,7 +50,7 @@ module.exports = {
         allowNull: false,
       },
       price: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(7,2),
         allowNull: false,
       },
       createdAt: {
